feat: [배성현] 사용자 정보 수신/저장 및 무료 사주 결과 생성 Python 재구현 [HM-BE-9, HM-BE-10]

+++ app/domains/user/infrastructure/orm/saju_result_orm.py
@@ -1,6 +1,6 @@
 from typing import Any
 
-from sqlalchemy import JSON, BigInteger, ForeignKey, UniqueConstraint
+from sqlalchemy import JSON, ForeignKey, Integer, UniqueConstraint
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.infrastructure.database.session import Base
@@ -11,6 +11,6 @@ class SajuResultORM(Base):
     __table_args__ = (UniqueConstraint("user_id", name="uq_saju_results_user_id"),)
 
     id: Mapped[int] = mapped_column(primary_key=True, autoincrement=True)
-    user_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("users.id"), nullable=False)
+    user_id: Mapped[int] = mapped_column(Integer, ForeignKey("users.id"), nullable=False)
     fortuneteller_request: Mapped[dict[str, Any]] = mapped_column(JSON, nullable=False)
     fortuneteller_response: Mapped[dict[str, Any]] = mapped_column(JSON, nullable=False)
